addressing converter cleanup as per #7 - repackaging
