+++ uploader/pom.xml
@@ -16,11 +16,14 @@
     <maven.compiler.target>1.8</maven.compiler.target>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <quarkus-plugin.version>1.0.1.Final</quarkus-plugin.version>
+    <quarkus.platform.version>1.1.1.Final</quarkus.platform.version>
+    <quarkus-plugin.version>1.1.1.Final</quarkus-plugin.version>
     <quarkus.platform.artifact-id>quarkus-universe-bom</quarkus.platform.artifact-id>
     <quarkus.platform.group-id>io.quarkus</quarkus.platform.group-id>
-    <quarkus.platform.version>1.0.1.Final</quarkus.platform.version>
+
     <surefire-plugin.version>2.22.1</surefire-plugin.version>
+    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+
   </properties>
   <dependencyManagement>
     <dependencies>
