+++ uploader/src/main/java/com/liquidlabs/logscape/uploader/FileMetaDataQueryService.java
@@ -1,0 +1,20 @@
+package com.liquidlabs.logscape.uploader;
+
+import java.util.List;
+
+public interface FileMetaDataQueryService {
+    void createTable();
+
+    void put(FileMeta fileMeta);
+
+    FileMeta find(String tenant, String filename);
+
+    //TODO: get content needs to be removed to projection providers (i.e. deletegate to appropriate projection layer)
+    byte[] get(String tenant, String filename);
+
+    FileMeta delete(String tenant, String filename);
+
+    List<FileMeta> query(String tenant, String filenamePart, String tagNamePart);
+
+    List<FileMeta> list();
+}
