+++ uploader/src/main/java/com/liquidlabs/logscape/uploader/Storage.java
@@ -1,0 +1,13 @@
+package com.liquidlabs.logscape.uploader;
+
+public interface Storage {
+    /**
+     * Store and Capture the Storage URL
+     * @param upload
+     * @param region
+     * @return
+     */
+    FileMeta upload(FileMeta upload, String region);
+
+    byte[] get(String storageUrl);
+}
