+++ uploader/src/main/java/com/liquidlabs/logscape/uploader/StorageIndexer.java
@@ -1,0 +1,19 @@
+package com.liquidlabs.logscape.uploader;
+
+public interface StorageIndexer {
+
+    /**
+     * Capture
+     * - uncompessed file size - capture if needed
+     * - compressed file size
+     * - eoln character
+     * - number of lines
+     * - time-from
+     * - time-to
+     * @param fileMeta
+     * @return
+     */
+    FileMeta enrichMeta(FileMeta fileMeta);
+
+    FileMeta index(FileMeta fileMeta, String cloudRegion);
+}
