+++ uploader/src/main/java/com/liquidlabs/logscape/uploader/Uploader.java
@@ -1,0 +1,5 @@
+package com.liquidlabs.logscape.uploader;
+
+public interface Uploader {
+    
+}
