+++ uploader/src/main/java/com/liquidlabs/logscape/uploader/fixture/FixturedStorageConverter.java
@@ -1,0 +1,12 @@
+package com.liquidlabs.logscape.uploader.fixture;
+
+import com.liquidlabs.logscape.uploader.Storage;
+import org.eclipse.microprofile.config.spi.Converter;
+
+public class FixturedStorageConverter implements Converter<Storage> {
+
+    @Override
+    public Storage convert(String s) {
+        return new FixturedStorageService();
+    }
+}
