--- uploader/src/main/java/com/logscapeng/uploader/Uploader.java
@@ -1,5 +1,0 @@
-package com.logscapeng.uploader;
-
-public interface Uploader {
-    
-}
