+++ uploader/src/main/resources/META-INF/services/org.eclipse.microprofile.config.spi.Converter
@@ -1,8 +1,8 @@
 # Note: run with quarkus:test will use fixtures quarkus:dev will use local services with remote storage
 # ./mvnw package -Pnative -Dquarkus.profile=prod-aws`
 # quarkus:prod - The default profile when not running in development or test mode
-com.logscapeng.uploader.aws.AwsStorageConverter
-com.logscapeng.uploader.aws.AwsMetaDataQueryConverter
-com.logscapeng.uploader.fixture.FixturedIndexConverter
+com.liquidlabs.logscape.uploader.aws.AwsStorageConverter
+com.liquidlabs.logscape.uploader.aws.AwsMetaDataQueryConverter
+com.liquidlabs.logscape.uploader.fixture.FixturedIndexConverter
