+++ uploader/src/main/resources/application.properties
@@ -37,4 +37,4 @@ quarkus.log.file.level=TRACE
 quarkus.log.file.format=%d{HH:mm:ss} %-5p [%c{2.}]] (%t) %s%e%n
 
 # custom loggers
-quarkus.log.category."com.logscapeng".level=INFO
+quarkus.log.category."com.liquidlabs.logscape".level=INFO
