+++ uploader/src/main/resources/log4j2.xml
@@ -11,7 +11,7 @@
         </Console>
     </Appenders>
     <Loggers>
-        <Logger name="com.logscapeng" level="trace" additivity="false">
+        <Logger name="com.liquidlabs.logscape" level="trace" additivity="false">
             <AppenderRef ref="Console"/>
         </Logger>
         <Root level="INFO">
