+++ uploader/src/test/resources/application.properties
@@ -28,4 +28,4 @@ quarkus.log.file.level=TRACE
 quarkus.log.file.format=%d{HH:mm:ss} %-5p [%c{2.}]] (%t) %s%e%n
 
 # custom loggers
-quarkus.log.category."com.logscapeng".level=DEBUG
+quarkus.log.category."com.liquidlabs.logscape".level=DEBUG
