+++ uploader/src/test/resources/log4j2.xml
@@ -15,7 +15,7 @@
             <AppenderRef ref="Console"/>
         </Logger>
 
-        <Logger name="com.logscapeng" level="trace" additivity="false">
+        <Logger name="com.liquidlabs.logscape" level="trace" additivity="false">
             <AppenderRef ref="Console"/>
         </Logger>
         <Root level="WARN">
